Allow Sphinx v2+ as a dependency
This project seems to be unmaintained so we can't expect the upstream to allow
more recent versions of Sphinx for us.

This relaxes the dependency requirement on Sphinx to allow sphinx-jsondomain to
be used with Sphinx v2+. This is necessary as Pipenv currently can't overrule
sub-dependency versions: https://github.com/pypa/pipenv/issues/1921.