Allow empty strings as propexample.
Previously, leaving a propexample with no argument would cause an exception, making
it impossible to give an empty string as an example. This should fix it. Exceptions
or weird results can probably still come up if one uses other data types and the code
looks kind of fragile, but it works for now.